[refactor]: spinner 추가 및 config localhost로 추가하여 data 주소 바꿔줌
