+++ src/config.js
@@ -3,3 +3,4 @@ export const REDIRECT_URI = 'http://localhost:3000/oauth/kakao';
 export const CLIENT_SECRET = 'TvTcmwjpl0VIqiYTsbyA6Ce9qBnMN67z';
 export const KAKAO_AUTH_URL = `https://kauth.kakao.com/oauth/authorize?client_id=${REST_API_KEY}&redirect_uri=${REDIRECT_URI}&response_type=code`;
 export const KAKAOINIT = '407ccf3a53942a0a9a43fa86a6e8590f';
+export const localhost = 'http://localhost:8000';
