+++ src/pages/Author/AuthorProfile/PostBox.js
@@ -1,10 +1,57 @@
-import React from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import styled, { css } from 'styled-components';
-
+import Spinner from '../../List/Spinner';
 import { useNavigate } from 'react-router-dom';
+import { localhost } from '../../../config';
 
-function PostBox({ lists, target }) {
+function PostBox({ userId }) {
   const navigate = useNavigate();
+  const target = useRef(null);
+  const [count, setCount] = useState(2);
+  const [spinner, setSpinner] = useState(true);
+  const [lists, setLists] = useState([]);
+  const token = localStorage.getItem('token');
+
+  useEffect(() => {
+    fetch(`${localhost}/list/profile/${userId}?page=1&pageSize=6`, {
+      method: 'GET',
+    })
+      .then(res => res.json())
+      .then(data => {
+        setLists(data);
+      });
+  }, [userId]);
+
+  const fetchData = async () => {
+    setTimeout(async () => {
+      await fetch(
+        `${localhost}/list/profile/${userId}?page=${count}&pageSize=6`
+      )
+        .then(res => res.json())
+        .then(data => {
+          if (data !== null) {
+            setLists(lists.concat(data));
+          } else {
+            setSpinner(false);
+          }
+        });
+      setCount(count + 1);
+    }, 700);
+  };
+  useEffect(() => {
+    let observer;
+    if (target.current) {
+      observer = new IntersectionObserver(handleObserver, { threshold: 0.4 });
+      observer.observe(target.current);
+    }
+    return () => observer && observer.disconnect();
+  }, [lists]);
+
+  const handleObserver = async ([entry], observer) => {
+    if (entry.isIntersecting) {
+      await fetchData();
+    }
+  };
   return (
     <PostContainer ref={target}>
       {lists === null ||
@@ -24,7 +71,8 @@ function PostBox({ lists, target }) {
             </Container>
           </Post>
         ))}
-      <div ref={target} style={{ border: '1px solid rgba(0,0,0,0)' }} />
+      <div ref={target} style={{ border: '1px solid transparent' }} />
+      <SpinnerWrapper>{spinner && <Spinner />}</SpinnerWrapper>
     </PostContainer>
   );
 }
@@ -103,4 +151,9 @@ const ImgBox = styled.div`
       `;
     }
   }}
+`;
+const SpinnerWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  padding: 15px 0;
 `;
