+++ src/pages/MyProfile/MyProfileLayout.js
@@ -3,13 +3,14 @@ import styled from 'styled-components';
 import Header from '../components/Header/Header';
 import MyProfileTop from './MyProfileTop';
 import MyProfileCenter from './MyProfileCenter';
+import { localhost } from '../../config';
 
 function MyProfileLayout() {
   const [profileData, setProfileData] = useState({});
   useEffect(() => {
     const token = localStorage.getItem('token');
 
-    fetch(`http://localhost:8000/user/myProfile`, {
+    fetch(`${localhost}/user/myProfile`, {
       method: 'GET',
       headers: { 'Content-Type': 'application/json', token: token },
     })
